voice_clone_video_synthesis/pyproject.toml
[project]
name = "voice-clone-video-synthesis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # IMPORTANT: Always run `uv lock` or `uv lock --upgrade` to resolve dependencies
    # and update the lockfile after editing anything below.
    # WARNING: Ensure that you don't have a local `uv.toml` which overrides PyPI
    # while generating the lockfile: https://github.com/astral-sh/uv/issues/15741
    "accelerate==1.8.1",
    "cn2an==0.5.22",
    "cython==3.0.7",
    "descript-audiotools==0.7.2",
    "einops>=0.8.1",
    "ffmpeg-python==0.2.0",
    "g2p-en==2.1.0",
    "jieba==0.42.1",
    "json5==0.10.0",
    "keras==2.9.0",
    "librosa==0.10.2.post1",
    "loguru>=0.7.3",
    "matplotlib==3.8.2",
    "modelscope==1.27.0",
    "munch==4.0.0",
    "ninja>=1.13.0",
    "numba==0.58.1",
    "numpy==1.26.2",
    "omegaconf>=2.3.0",
    "opencv-python==4.9.0.80",
    "pandas==2.3.2",
    "safetensors==0.5.2",
    "sentencepiece>=0.2.1",
    "tensorboard==2.9.1",
    "textstat>=0.7.10",
    "tokenizers==0.21.0",
    "torch==2.8.*",
    "torchaudio==2.8.*",
    "tqdm>=4.67.1",
    "transformers==4.52.1",
    # Use "wetext" on Windows/Mac, otherwise "WeTextProcessing" on Linux.
    "wetext>=0.0.9; sys_platform != 'linux'",
    "WeTextProcessing; sys_platform == 'linux'",
]
[tool.uv]
index-url = "https://mirrors.aliyun.com/pypi/simple"
